requirements.txt
numpy==1.26.4
sympy==1.13.1
Flask==2.1.3
pillow==10.4.0
